+++ src/main/java/org/sopt/common/exception/GlobalExceptionHandler.java
@@ -1,5 +1,6 @@
 package org.sopt.common.exception;
 
+import org.sopt.common.response.ErrorResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
